Pass PagingViewController as argument to delegate
PagingViewControllerDelegate should pass in the
PagingViewController that's calling it. In order to make the
PagingItems helpers testable without having to pass in the
PagingViewController, we create a separate protocol that should
only return the width of a given paging items. This way the
helpers has no knowledge of the PagingViewController. This also
means that we no longer need the default delegate to return any
static width options, as this can be handled directly in the
protocol implementation.