Move global PagingItems helpers into extension
Adding a default implemenation for the PagingItemsPresentable
protocol means we don't have to pass in the presenable object for
every function. This also adds methods for returning the paging
items before or after any given items, allowing us to remove the
data source dependency. This means less things to pass in and
less things the tests need to have knowledge about.