Refactor the paging state machine
This refactor removes some unnecessary states and events on the state
machine and makes it much easier to understand what’s going on. The
previous implementation had a bunch of bugs where the states could get
out of sync if the user swiped a lot back and forth. This rewrite fixes
these issues and also includes tests for all of these edge cases. This
also removes the event and state observers in favour of configurable
blocks for the two events we care about.