Account for menuInsets when using sizeToFit
Previously, if the user specified non-zero left and/or right `menuInsets` & used `.SizeToFit` for `menuItemSize`, it would cause the menu to be horizontal scrollable, even if the min width + insets fit inside the available space.

Now the fitting check calculation includes the inset, and also subtracts the inset before dividing the available space, ensuring that when possible, the menu will not scroll.