Add progress value to menu items
Define a custom UICollectionViewLayoutAttributes subclass that we can
set a progress value on. The progress value is based on how much the
content is scrolled. In the cells we can override applyLayoutAttributes
and update the cell based on the progress of the transition.

In the default PagingTitleCell, we use this progress value to
interpolate between the selected and default text color.