Scroll menu items alongside content
Updates the scroll offset based on how much the content has scrolled.
This makes the menu items transition smoothly as you scroll. The old
behaviour is available as an option on the menuTransition property.

Getting the menu items to scroll alongside the content requires us to
calculate the distance between the current and upcoming paging item,
and then update the content offset based on the transition progress. We
also need to keep track of the initial content offset, so we create a
new PagingTransition struct that can store all this data.