Fix calculation for transition distance
When swiping to pages on the left or right edge we previously just took
the min and max value of the content offset to make sure it didn’t move
too far. This caused the transition to suddenly stop instead of
smoothly transitioning to the correct place.

This updated checks if the new content offset will be near any of the
edges and calculate how much it will overshoot. Then we can just
subtract that value from the original distance to get the correct
transition.