Fix issue with selecting paging item
Adding support for selecting paging items before viewDidAppear (#32)
introduced a bug when navigating back and fourth to the paging view,
for instance when using a tab bar.

EMPageViewController does not support calling selectViewController with
the same view controller that is currently selected. When doing that it
selects the next view controller and causes corrupted menu items. To
fix this we just return if you select the same view controller twice.