Update EMPageViewController delegate method
When scrolling away from a page near an edge, it should trigger the
`isScrollingFrom:` delegate method with nil. This is a change that was
added to EMPageViewController so that the paging indicator will
overshoot when dragging away from the edges.

When updating the EMPageViewController version in #36, this fix was not
included. Should really move this to Carthage.