Remove view controller selection in viewDidAppear
When viewDidAppear is called it triggers viewControllerForPagingItem
data source method. This can potentially create a new instance of the
selected view controller each time viewDidAppear is called.

Calling selectViewController in viewDidAppear was added to allow
selectPagingItem to be called before viewDidAppear (#32). Turns out
it’s sufficient to only call selectViewController when calling
selectPagingItem.