Enable scrolling in the menu items header
Since Parchment supports infinite data sources we only display the
items that are nessesary to cover the bounds. When enabling scrolling
in the collection view we need to reload the data as you are scrolling
near an edge so that it looks like it is a continous scroll.

By default, scrolling is now enabled in in header. It is still
possible to just use the swipe gesture by setting the new
`menuGesture: PagingMenuGesture` option to `.swipe`.