Require PagingItem to conform to Hashable and Comparable
Conforming to Hashable means we can put the paging items into a
dictionary that makes it easier to diff the changes between to sets of
paging items. We can also store the paging items in a Set which makes
easier and more performant to get the union between two datasets. This
change will break any existing PagingItem implementation.