Write custom collection view layout
Replaces `UICollectionViewFlowLayout` with a custom collection view
layout implementation. This allows us to move much of the code from
the `PagingViewController` into the layout were it should belong. This
also makes it much easier to write tests for the actual layout code.

The new layout uses a custom invalidation context that allow us to
only invalidate certain parts without having to recalculate the entire
layout. It also moves the transition object from the state machine and
into the layout. Using a custom layout also fixes multiple issues with
`UICollectionViewFlowLayout` and makes it easier to debug as we don't
need to worry about the internal state of the flow layout.

This commit also refactors the way we calculate the content offset
after reloading items. Instead of using the weird
PagingItemsPresentable protocol we create a separate struct that is
only responsible for diffing two paging item datasets. We can then
update the offset based on the width of the added/removed items.