Fix offset calculation when cells have different width
When using different widths for each cell the offset calculation did
not return the correct value. We need to read from the old layout
attributes when getting the width from removed items as they may no
longer be in the layout. We also need to manually call prepare() on
the layout to make sure the new layout attributes are updated before
we calculate the offset.