Fix bug where selecting items was not working
Scrolling outside from either of the edges disabled interaction with
the menu items. This was caused by not resetting the state back to
.selected after the transition completed. Now we trigger a new event
when a transition has cancelled. We also rename the previous event
called cancelScrolling to transitionSize and use the cancelScrolling
event to reset the state back to selected.