Fix selecting paging items before view appears
When using FixedPagingViewController, calling selectPagingItem in
viewDidLoad did not update the selected paging item. The conditional
check inside selectPagingItem did not account for situations where the
PagingDataStructure object was not initialized (which only happens
after the view appears), preventing the selection to update.

We also need to account for the fact that EMPageViewController will
trigger its scrolling delegate event even if the view has not appeared,
causing the wrong initial paging item.