Fix menu item transition issue with safe area insets
When swiping between items on the iPhone X the distance calculation did
not account for the safe area insets, causing the content offset to
jump. Setting the contentInsetAdjustmentBehavior to .never and applying
the safe area insets manually fixes this issue.

We also need to move the initial render into viewDidLayoutSubviews, as
the safeAreaInsets property is not  updated in viewWillAppear.