Allow setting different width for selected cell
The widthForPagingItem delegate now has a new argument called
isSelected which can be used to return a different width for the
currently selected cell. Parchment will then tween between the selected
width and the default width when swiping between items.

If you implement the width delegate, the collection view layout will
invalidate all the layout attributes as you are scrolling. The menu
item sizes are cached so they are not re-calculated all the time.