Replace PagingOptions with property assignments
This change introduces a breaking change. All configuration is now
moved to properties on the PagingViewController class instead of having
to pass in an instance conforming to the PagingOptions protocol.

This reduces the boilerplate of having to create a separate options
struct when you just need to override a single value. It also means
you can change the options after the PagingViewController has been
initialized. Lastly, this change also means we can introduce better
support for customizations through Storyboard in the future.