Update API for paging data source
Introduces a new data source protocol that makes it easier to setup a
custom data source if you have a fixed number of view controllers. To
use the new data source, you just need to return the total number of
view controllers as well as the view controller and PagingItem for a
given index. It’s still possible to setup an infinite data source, but
the protocol has been renamed to PagingViewControllerInfiniteDataSource
and moved into a property called infiniteDataSource.

This also makes some changes to FixedPagingViewController by using the
new data source and replacing the ViewControllerItem with
PagingIndexItem. PagingIndexItem now just stores the index and title,
instead of holding a reference to the view controller.