Update API for delegate methods
Adds three new delegate methods to the PagingViewControllerDelegate
protocol that is called before, during and after the user navigates to
another view controller.

To determine whether someone was using dynamic widths using
widthForPagingItem, we previously just checked if the delegate was
implemented. This will no longer work as we have multiple delegate
methods, and we cannot prefix it with @objc optional as we are using
generic parameters. Therefore, the return type for widthForPagingItem
is now optional with a default implementation that returns nil.

With these new delegate methods we can remove the
FixedPagingViewControllerDelegate protocol. The new delegate methods
does not include the index for the paging items, but if you are using
FixedPagingViewController this will be available as a property on the
PagingIndexItem instance.