Fix layout calculation when using transforms
When applying transforms to the menu cells the layout calculations
would give the wrong result. This happens because the frame property is
just a computed value that’s based on the center, bounds and transform
properties. We therefore need to read from bounds and center properties
directly to prevent the layout calculation to be affected by any
transforms.