Allow subclassing the collection view layout
Adds a property called menuLayoutClass that allows you to override the
PagingCollectionViewLayout class type. This means you can provide your
own subclass of the layout.

Since the collection view layout type will be set after the
PagingViewController has been initialized we need to wait to initialize
the collection view layout until the view has loaded. This means both
the collectionViewLayout and the collectionView are now optional. This
also means we need to update the initializer for the PagingView. It now
takes the collectionView and the pageView as arguments in the configure
function, which is called after the view has loaded.