Fix performance issue with updating content offset
When updating the content offset while scrolling relatively fast the
menu items would sometimes lag behind. This was mostly visible in
performance heavy situations, like when using the width delegate
alongside some resource heavy view controllers. Moving the content
offset invalidation out of the collection view layout fixes this issue.

With these changes, the invalidation context is only used to update
the layout attributes, and not to calculate the distance to the next
item or to update the content offset. The PagingViewController is now
responsible for updating the content offset, while the distance
calculation is moved into its own separate struct. This also makes the
distance calculation much easier to test.