Change collection view properties to non-optional
When adding the menuLayoutClass property the collectionView and
collectionViewLayout properties was changed to optional since the
layout type can change after initialization.

Instead of changing the properties to optional we can initialize the
collection view with the default layout type when we create the
PagingViewController and then only re-create the layout if the
menuLayoutClass property changes. This means we can keep the collection
view and collection view layout non-optional, and it will also allow
you to change the layout type after the view has loaded.