Fix issue with selecting items programatically
After version 1.0, selecting items programatically before the view has
appeared no longer works. This is caused by the fact that we receive
delegate methods from the page view controller when setting the
initial view controller before the view has appeared. To fix this we
wait until the view is ready before setting the page view delegate. We
also need to add a new PagingEvent that we can use to select items
before the page view delegate is configured.