Fix issue with menu items being clipped
When using Parchment without a UINavigationController the menu items
would get clipped. This happened because we inset the menu items based
on the safe area insets, which gets a larger values when using without
a navigation controller. We only need to account for the safe area
insets horizontally so we can just remove the top and bottom insets.