Fix layout bug with cell heights when using width delegate
When using the width delegate, the collection view layout would use
the computed property menuHeight when calculating the frame of the
cell. This property includes the menu insets, causing the cells to get
too large when using the menuInsets property alongside the width
delegate. Using the value of menuItemSize.height fixes this issue.