Fix issues with reloading data
The current implementation of reloadData was causing unexpected
results and would sometimes crash the application. This happened
because we were relying on the existing select method to reload
data. Instead we create a separate method that makes sure to reload
both the collection view and page view controller to the new state.