Prevent unnecessary view controller initializations
When selecting menu items we always called “selectViewController” on
the page view controller. This will initialize new view controllers
around the selected item, even if they already exist. Instead we can
check if the selected item is the next or previous one and just
animate directly to that without initializing a new one.