Remove generics from PagingViewController
Removing the generic type on PagingViewController will clean up the
API quite a lot. Using generics does not really give us any benefits,
as you need to cast the type in the delegate and data sources anyway.

This commit also renames PagingIndexItem to PagingTitleItem, and
updates the data source protocols to be a bit more Swifty.