Fix issue changing selection when view is not visible
When calling select(pagingItem:) while the view controller was not
visible, the menu items would get out of sync when navigating
back. This happened because the state never changed to .scrolling and
therefore the .cancelScrolling event was triggered. To fix this we
need to check that the state is actually .scrolling before calling the
.cancelScrolling event.