Fix regression causing indicator not to animate
When the width of the menu items are less than the bounds width the
indicator does not animate while scrolling. This issue was introduced
in the 2.0 refactor. Fixed by ensuring we invalidate the collection
view layout regardless of the menu item size (it even says we need
this in the comments, but it was messed up during the refactor).