Combine all example targets into single target
Having the examples in separate targets makes it quite hard to make
sure that everything is working as you have to compile each target
separately. Combining them into one also makes it easier to reuse code
between the examples, and it's easier for people checking out the
framework to see what is possible with it.