Add delegate method when selecting item /w @yosshi4486
Adds a new delegate method called didSelectItem that is called when
the user selects an item in the menu. This is different from the
existing didScrollToItem delegate method, as this will be called even
if you are already on that page. This method will be called before the
collection view starts scrolling to the new item.