Add support calculating vertical distances
PagingDistance is responsible for calculating the distance that the
collection view has to travel when selecting an item. This is used to
ensure a smooth transition. This commit refactors the calculation to
work with both vertical and horizontal layouts and introduces unit
tests for the most of the possible cases.